requirements.txt
g4f~=0.0.3.1
discord.py~=2.3.2
PyExecJs~=1.5.1